+++ src/main/java/com/microsoft/azure/datalake/store/ADLFileInputStream.java
@@ -28,17 +28,6 @@ import java.util.UUID;
  *
  */
 public class ADLFileInputStream extends InputStream {
-
-    /*
-    Hadoop interfaces not implemented: (Remove comment after first Code Review from Vishwajeet)
-     - PositionedReadable: can be done in the shim, using Core.open directly
-     - ByteBufferReadable: can be done in the shim, using Core.open directly
-     - HasFileDescriptor: seems orthogonal to this class's functionality
-     - CanSetDropBehind: seems like a synonym of the unbuffer() method (already implemented in this class)
-     - HasEnhancedByteBufferAccess: Not sure why this even is in the FSDataInputStream - seems orthogonal
-     - CanSetReadahead/CanSetDropbehind - not that hard to do, but not sure who uses it. Will do if needed.
-     */
-
     private static final Logger log = LoggerFactory.getLogger("com.microsoft.azure.datalake.store.ADLFileInputStream");
 
     private final String filename;
@@ -52,7 +41,7 @@ public class ADLFileInputStream extends InputStream {
     private long fCursor = 0;  // cursor of buffer within file - offset of next byte to read from remote server
     private int bCursor = 0;   // cursor of read within buffer - offset of next byte to be returned from buffer
     private int limit = 0;     // offset of next byte to be read into buffer from service (i.e., upper marker+1
-                               //                                                      of valid bytes in buffer)
+    //                                                      of valid bytes in buffer)
     private boolean streamClosed = false;
 
 
@@ -67,63 +56,12 @@ public class ADLFileInputStream extends InputStream {
         }
     }
 
-    /**
-     * Read upto the specified number of bytes, from a given
-     * position within a file, and return the number of bytes read. This does not
-     * change the current offset of a file.
-     *
-     * @param position position in file to read from
-     * @param b  byte[] buffer to read into
-     * @param offset offset into the byte buffer at which to read the data into
-     * @param length number of bytes to read
-     * @return the number of bytes actually read, which could be less than the bytes requested. If the {@code position}
-     *         is at or after end of file, then -1 is returned.
-     * @throws IOException thrown if there is an error in reading
-     */
-    public int read(long position, byte[] b, int offset, int length)
-            throws IOException {
-        if (streamClosed) throw new IOException("attempting to read from a closed stream");
-        if (position < 0) throw new IllegalArgumentException("attempting to read from negative offset");
-        if (position >= directoryEntry.length) return -1;  // Hadoop prefers -1 to EOFException
-        if (b == null) throw new IllegalArgumentException("null byte array passed in to read() method");
-        if (offset >= b.length) throw new IllegalArgumentException("offset greater than length of array");
-        if (length < 0 || length > (b.length - offset))
-            throw new IllegalArgumentException("requested read length is more than will fit after requested offset in buffer");
-
-        if (log.isTraceEnabled()) {
-            log.trace("ADLFileInputStream positioned read() - at offset {} using client {} from file {}", position, client.getClientId(), filename);
-        }
-
-        // make server call to get more data
-        RequestOptions opts = new RequestOptions();
-        opts.retryPolicy = new ExponentialBackoffPolicy();
-        OperationResponse resp = new OperationResponse();
-        InputStream inStream = Core.open(filename, position, length, sessionId, client, opts, resp);
-        if (!resp.successful) throw client.getExceptionFromResponse(resp, "Error reading from file " + filename);
-        if (resp.responseContentLength == 0 && !resp.responseChunked) return 0;  //Got nothing
-        int bytesRead;
-        int totalBytesRead = 0;
-        try {
-            do {
-                bytesRead = inStream.read(b, offset + totalBytesRead, length - totalBytesRead);
-                if (bytesRead > 0) { // if not EOF of the Core.open's stream
-                    totalBytesRead += bytesRead;
-                }
-            } while (bytesRead >= 0 && totalBytesRead < length);
-        } catch (IOException ex) {
-            throw new ADLException("Error reading data from response stream in positioned read() for file " + filename, ex);
-        } finally {
-            inStream.close();
-        }
-        return totalBytesRead;
-    }
-
     @Override
     public int read() throws IOException {
         byte[] b = new byte[1];
         int i = read(b, 0, 1);
         if (i<0) return i;
-            else return (b[0] & 0xFF);
+        else return (b[0] & 0xFF);
     }
 
     @Override
@@ -178,10 +116,6 @@ public class ADLFileInputStream extends InputStream {
         if (bCursor < limit) return 0; //if there's still unread data in the buffer then dont overwrite it
         if (fCursor >= directoryEntry.length) return -1; // At or past end of file
 
-        if (log.isTraceEnabled()) {
-            log.trace("ADLFileInputStream.readFromService() - at offset {} using client {} from file {}", getPos(), client.getClientId(), filename);
-        }
-
         if (directoryEntry.length <= blocksize)
             return slurpFullFile();
 
@@ -189,48 +123,11 @@ public class ADLFileInputStream extends InputStream {
         bCursor = 0;
         limit = 0;
         if (buffer == null) buffer = new byte[blocksize];
-        byte[] junkbuffer = new byte[16*1024];
 
-        // make server call to get more data
-        int loopcount = 0;
-        int totalBytesRead = 0;
-        while (totalBytesRead == 0) {
-            RequestOptions opts = new RequestOptions();
-            opts.retryPolicy = new ExponentialBackoffPolicy();
-            OperationResponse resp = new OperationResponse();
-            InputStream str = Core.open(filename, fCursor, blocksize, sessionId, client, opts, resp);
-            if (!resp.successful) throw client.getExceptionFromResponse(resp, "Error reading from file " + filename);
-            if (resp.responseContentLength == 0 && !resp.responseChunked) return 0;  //Got nothing
-            int bytesRead;
-            try {
-                do {
-                    bytesRead = str.read(buffer, limit, blocksize - limit);
-                    if (bytesRead > 0) { // if not EOF of the Core.open's stream
-                        limit += bytesRead;
-                        fCursor += bytesRead;
-                        totalBytesRead += bytesRead;
-                    }
-                } while (bytesRead >= 0 && limit < blocksize);
-                if (bytesRead >= 0) {  // read to EOF on the stream, so connection can be reused
-                    while (str.read(junkbuffer, 0, junkbuffer.length)>=0); // read and consume rest of stream, if any remains
-                }
-                if (totalBytesRead == 0) {
-                    loopcount++;
-                    if (loopcount >= 5) throw client.getExceptionFromResponse(resp, "No data read data after " + loopcount + " attempts from response stream for file " + filename);
-                }
-            } catch (Exception ex) {
-                if (totalBytesRead > 0) {
-                    // if anything has been read, then forget about the exception and just return what we got
-                    return totalBytesRead;
-                } else {
-                    loopcount++;
-                    if (loopcount >= 5) throw client.getExceptionFromResponse(resp, "Error reading data after " + loopcount + " attempts from response stream for file " + filename);
-                }
-            } finally {
-                if (str != null) str.close();
-            }
-        }
-        return totalBytesRead;
+        int bytesRead = readInternal(fCursor, buffer, 0, blocksize);
+        limit += bytesRead;
+        fCursor += bytesRead;
+        return bytesRead;
     }
 
 
@@ -245,50 +142,89 @@ public class ADLFileInputStream extends InputStream {
             log.trace("ADLFileInputStream.slurpFullFile() - using client {} from file {}. At offset {}", client.getClientId(), filename, getPos());
         }
 
-        //reset buffer to initial state - i.e., throw away existing data
-        if (getPos() > Integer.MAX_VALUE)  throw new IOException("Attempting to slurp file that is too big: " + filename);
-
         if (buffer == null) {
             blocksize = (int) directoryEntry.length;
             buffer = new byte[blocksize];
         }
 
-        bCursor = (int) getPos();
+        //reset buffer to initial state - i.e., throw away existing data
+        bCursor = (int) getPos();  // preserve current file offset (may not be 0 if app did a seek before first read)
         limit = 0;
         fCursor = 0;  // read from beginning
         int loopCount = 0;
-        byte[] junkbuffer = new byte[16*1024];
 
-        while (fCursor < directoryEntry.length) {   // if one OPEN request doesnt get full file, then read again at fCursor
-            RequestOptions opts = new RequestOptions();
-            opts.retryPolicy = new ExponentialBackoffPolicy();
-            OperationResponse resp = new OperationResponse();
-            InputStream str = Core.open(filename, fCursor, blocksize, sessionId, client, opts, resp);
-            if (!resp.successful) throw client.getExceptionFromResponse(resp, "Error reading from file " + filename);
-            if (resp.responseContentLength == 0 && !resp.responseChunked) continue;  //Got nothing, try again
-            int bytesRead;
-            try {
-                do {
-                    bytesRead = str.read(buffer, limit, blocksize - limit);
-                    if (bytesRead > 0) { // if not EOF of the Core.open's stream
-                        limit += bytesRead;
-                        fCursor += bytesRead;
-                    }
-                } while (bytesRead >= 0 && limit < blocksize);
-                if (bytesRead >= 0) {  // read to EOF on the stream, so connection can be reused
-                    while (str.read(junkbuffer, 0, junkbuffer.length)>=0); // read and consume rest of stream, if any remains
-                }
-            } catch (Exception ex) {
-                // swallow, since we will consume however many bytes we got and then retry for the rest of the bytes
-            } finally {
-                if (str!=null) str.close();
-            }
+        // if one OPEN request doesnt get full file, then read again at fCursor
+        while (fCursor < directoryEntry.length) {
+            int bytesRead = readInternal(fCursor, buffer, limit, blocksize - limit);
+            limit += bytesRead;
+            fCursor += bytesRead;
+
             // just to be defensive against infinite loops
             loopCount++;
-            if (loopCount > 10) { throw new IOException("Too many attempts in reading whole file " + filename); }
+            if (loopCount >= 10) { throw new IOException("Too many attempts in reading whole file " + filename); }
         }
         return fCursor;
     }
+
+    /**
+     * Read upto the specified number of bytes, from a given
+     * position within a file, and return the number of bytes read. This does not
+     * change the current offset of a file.
+     *
+     * @param position position in file to read from
+     * @param b  byte[] buffer to read into
+     * @param offset offset into the byte buffer at which to read the data into
+     * @param length number of bytes to read
+     * @return the number of bytes actually read, which could be less than the bytes requested. If the {@code position}
+     *         is at or after end of file, then -1 is returned.
+     * @throws IOException thrown if there is an error in reading
+     */
+    public int read(long position, byte[] b, int offset, int length)
+            throws IOException {
+        if (streamClosed) throw new IOException("attempting to read from a closed stream");
+        if (log.isTraceEnabled()) {
+            log.trace("ADLFileInputStream positioned read() - at offset {} using client {} from file {}", position, client.getClientId(), filename);
+        }
+
+        return readInternal(position, b, offset, length);
+    }
+
+    private int readInternal(long position, byte[] b, int offset, int length) throws IOException {
+        if (position < 0) throw new IllegalArgumentException("attempting to read from negative offset");
+        if (position >= directoryEntry.length) return -1;  // Hadoop prefers -1 to EOFException
+        if (b == null) throw new IllegalArgumentException("null byte array passed in to read() method");
+        if (offset >= b.length) throw new IllegalArgumentException("offset greater than length of array");
+        if (length < 0) throw new IllegalArgumentException("requested read length is less than zero");
+        if (length > (b.length - offset))
+            throw new IllegalArgumentException("requested read length is more than will fit after requested offset in buffer");
+
+        byte[] junkbuffer = new byte[16*1024];
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new ExponentialBackoffPolicy();
+        OperationResponse resp = new OperationResponse();
+        InputStream inStream = Core.open(filename, position, length, sessionId, client, opts, resp);
+        if (!resp.successful) throw client.getExceptionFromResponse(resp, "Error reading from file " + filename);
+        if (resp.responseContentLength == 0 && !resp.responseChunked) return 0;  //Got nothing
+        int bytesRead;
+        int totalBytesRead = 0;
+        try {
+            do {
+                bytesRead = inStream.read(b, offset + totalBytesRead, length - totalBytesRead);
+                if (bytesRead > 0) { // if not EOF of the Core.open's stream
+                    totalBytesRead += bytesRead;
+                }
+            } while (bytesRead >= 0 && totalBytesRead < length);
+            if (bytesRead >= 0) {  // read to EOF on the stream, so connection can be reused
+                while (inStream.read(junkbuffer, 0, junkbuffer.length)>=0); // read and consume rest of stream, if any remains
+            }
+        } catch (IOException ex) {
+            throw new ADLException("Error reading data from response stream in positioned read() for file " + filename, ex);
+        } finally {
+            if (inStream != null) inStream.close();
+        }
+        return totalBytesRead;
+    }
+
 
     /**
      * Seek to given position in stream.
@@ -422,7 +358,7 @@ public class ADLFileInputStream extends InputStream {
      */
     @Override
     public synchronized void mark(int readlimit) {
-      throw new UnsupportedOperationException("mark()/reset() not supported on this stream");
+        throw new UnsupportedOperationException("mark()/reset() not supported on this stream");
     }
 
     /**